+++ test/test.js
@@ -46,6 +46,40 @@ describe('csurf', function () {
     });
   });
 
+  it('should work in csrf-token header', function(done) {
+    var server = createServer()
+
+    request(server)
+    .get('/')
+    .expect(200, function (err, res) {
+      if (err) return done(err)
+      var token = res.text;
+
+      request(server)
+      .post('/')
+      .set('Cookie', cookies(res))
+      .set('csrf-token', token)
+      .expect(200, done)
+    });
+  });
+
+  it('should work in xsrf-token header', function(done) {
+    var server = createServer()
+
+    request(server)
+    .get('/')
+    .expect(200, function (err, res) {
+      if (err) return done(err)
+      var token = res.text;
+
+      request(server)
+      .post('/')
+      .set('Cookie', cookies(res))
+      .set('xsrf-token', token)
+      .expect(200, done)
+    });
+  });
+
   it('should work in x-csrf-token header', function(done) {
     var server = createServer()